feat(Tests.System): Use class fixture for container lifetime
